Split BYOS logic and example into two packages and add classes to move schema specific code out of logic files.
