+++ src/main/kotlin/example/ByosApplication.kt
@@ -1,4 +1,4 @@
-package byos
+package example
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
